refactor: keep topic scan media manifest out of stdout
